requirements.txt
python-telegram-bot==20.8
asyncpg
stripe
python-dotenv